+++ src/lib/util/constants.js
@@ -23,7 +23,7 @@ exports.DEFAULTS = {
 		},
 		disabledCorePieces: [],
 		language: 'en-US',
-		prefix: '!',
+		prefix: '',
 		preserveSettings: true,
 		readyMessage: (client) => `Successfully initialized. Ready to serve ${client.guilds.size} guild${client.guilds.size === 1 ? '' : 's'}.`,
 		typing: false,