+++ src/monitors/commandHandler.js
@@ -55,14 +55,14 @@ module.exports = class extends Monitor {
 
 	getPrefix(message) {
 		if (this.prefixMention.test(message.content)) return { length: this.nick.test(message.content) ? this.prefixMentionLength + 1 : this.prefixMentionLength, regex: this.prefixMention };
-		if (message.guildSettings.disableNaturalPrefix !== true && this.client.options.regexPrefix) {
+		if (!message.guildSettings.disableNaturalPrefix && this.client.options.regexPrefix) {
 			const results = this.client.options.regexPrefix.exec(message.content);
 			if (results) return { length: results[0].length, regex: this.client.options.regexPrefix };
 		}
-		const prefix = message.guildSettings.prefix || this.client.options.prefix;
+		const { prefix } = message.guildSettings;
 		if (Array.isArray(prefix)) {
-			for (let i = prefix.length - 1; i >= 0; i--) {
-				const testingPrefix = this.prefixes.get(prefix[i]) || this.generateNewPrefix(prefix[i]);
+			for (const prf of prefix) {
+				const testingPrefix = this.prefixes.get(prf) || this.generateNewPrefix(prf);
 				if (testingPrefix.regex.test(message.content)) return testingPrefix;
 			}
 		} else if (prefix) {